Move docker.genericImage to a separate skopeo/image subpackage
... making image.FromSource a public, stable, API.
