+++ signature/policy_eval_signedby_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/projectatomic/skopeo/directory"
-	"github.com/projectatomic/skopeo/docker"
+	"github.com/projectatomic/skopeo/image"
 	"github.com/projectatomic/skopeo/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -15,7 +15,7 @@ import (
 
 // dirImageMock returns a types.Image for a directory, claiming a specified intendedDockerReference.
 func dirImageMock(dir, intendedDockerReference string) types.Image {
-	return docker.GenericImageFromSource(&dirImageSourceMock{
+	return image.FromSource(&dirImageSourceMock{
 		ImageSource:             directory.NewDirImageSource(dir),
 		intendedDockerReference: intendedDockerReference,
 	})
